refactor(gateway): split responses middleware + fix logprobs intent recording
Two related tidy-ups on the middleware layer added in the previous commit.

1. Split middleware.py by concern:
   - ``middleware.py`` keeps ``SessionRoutingMiddleware`` plus the shared
     ``_read_body`` / ``_body_replay`` ASGI helpers (183 lines).
   - New ``responses_middleware.py`` holds ``ResponsesAdapterMiddleware``,
     ``_SSETranslatingSend``, and the ``_RESP_PATH_RE`` regex (235 lines).
   - ``server.py`` import path updated accordingly.
   Rationale: session routing and Responses protocol translation are two
   independent concerns — one cares about URL shape and sampling params,
   the other about API format translation. Bundling them into a single
   file made ``middleware.py`` grow to 407 lines and forced the module
   docstring to describe two unrelated middleware classes.

2. Fix a latent logprobs-intent recording bug:
   ``SessionRoutingMiddleware`` L102 unconditionally overwrites
   ``scope.state["originally_requested_logprobs"]`` by reading the Chat
   Completions ``logprobs`` field. For requests that arrived via the
   Responses API, the ``include`` list carries this intent — but by the
   time SessionRouting runs, the ResponsesAdapterMiddleware has already
   translated the body, and the Chat-shape body doesn't carry it.
   Result: a hypothetical Responses client asking for logprobs via
   ``include: [message.output_text.logprobs]`` would have them stripped
   before returning (bug).

   Fix respects the layering: ResponsesAdapterMiddleware records intent
   from its own convention (``include``) onto scope state before
   delegating downstream; SessionRoutingMiddleware wraps its own write
   in an ``if key not in state`` guard so it only falls back to Chat
   convention when no outer middleware has already recorded. Each
   middleware stays scoped to its own protocol.

Verified: 269 gateway unit + 58 harness tests pass. Codex CLI itself
doesn't currently request logprobs, so the fix is upstream-review defense
rather than a hot fix.

Known gap (Phase G): the training path's re-forward via
``verl_engine.process_image_data(messages)`` needs image_url content parts
in ``trace.messages``. Turn 2+ cumulative requests hit ``/v1/completions``
and produce empty ``trace.messages`` (their multimodal state lives in
``raw_request["multi_modal_data"]``). Propagating it through
``TraceRecord`` → ``Step`` → ``ModelOutput`` is out of Phase D scope and
tracked as follow-on training-side work.
